Make toolshelf.py write to the `source`d result file directly.

+++ toolshelf.sh
@@ -1,7 +1,9 @@
 # `source`d by the alias `toolshelf` which `bootstrap-toolshelf.sh` adds to
 # your `.bashrc`
 
-F=$TOOLSHELF/tmp-toolshelf-result.sh
-python $TOOLSHELF/toolshelf/toolshelf.py $* >$F
-source $F
-rm -f $F
+F=$TOOLSHELF/toolshelf/tmp-toolshelf-result.sh
+python $TOOLSHELF/toolshelf/toolshelf.py $*
+if [ -e $F ]; then
+    source $F
+    rm -f $F
+fi
